+++ Changes.rst
@@ -9,6 +9,8 @@ Unreleased Changes
 * `~HTTPConnection.send_request` now returns a `HTTPResponse` instance
   instead of a tuple.
 
+* The `!HTTPConnection.get_current_response` method has been removed.  
+
 Release 1.0 (2013-07-13)
 ========================